+sfbay.sunrisemovement.org, - make www-data owner
I'm going to try to clone git repositories with the correct group+owner by becoming www-data, rather than having a handler go through and recursively set the owner on everything, which appears to not be quite idempotent.

Also adding sfbay.sunrisemovement.org since everyone was happy with the .dev version, and updating the version of Caddy v2.