discard unnecessary commented code, update Hugo
We no longer need to provide or even use the root password, instead we're relying on SSH keys to log into the server immediately after creation.

Wheel group appears to be legacy, sudo seems to be the way to go today.

The "force change password" handler was merely moved later in the file.