handcoded RI PHP site now works
SFbay now uses Hugo instead of being handcoded, so we'll deal with that in the Hugo tasks section.

This should bring us roughly to feature parity with the main branch, since we never had Hugo sites working properly anyway, so I'm gonna make the "arch" branch the main branch. Maybe I'll back up the old branch for historical purposes... or maybe not.