install stable release of Caddy
Updated service file based on official service file at https://github.com/caddyserver/dist/tree/master/init

Right now we have Ubuntu using /usr/bin/caddy because we're using an official package, while Arch is still running a manually installed package in /usr/loca/bin/

Changing the location of the Caddy binary in the caddy.service file seems to require a daemon-reload AND a restart of Caddy.

The current code may cause some unnecessary restarts, but I'm not sure how to tell when a restart is necessary or when a simple reload would do the job.

In case you're wondering, @zarex, this code appears to work!