sloppy change storage location
Info about what I was trying to fix with this commit: https://caddy.community/t/v2-cant-use-local-ca-certs-storage-misconfigured/7986/

This embarassing commit resulted in two working web servers after I messed with the server manually, but I'm not confident this code did what it was supposed to. E.g. I thought I copied the old storage to the new location, but Caddy still seemed to fetch new certs anyway according to the logs.

I'm also not sure "caddy trust" does what I think it does, since I thought it would get rid of the "http : user NOT in sudoers" error, and it doesn't seem to have done so.

Since the new storage location is the default, which is based on the web server user's home directory, I changed the autosave and storage dirs to include home dir variables.