fixed linode_destroy, switch to Ubuntu
linode_destroy works now, but we have to get rid of the stupid practice of importing variables from a separate role.

I think we should use Ubuntu on the stable server and perhaps Arch on a testing server? Sadly the Caddy package on Arch is still not up to date.