switch to caddy-bin, eliminate favicon requests
We're switching from the caddy2 AUR package to caddy-bin, because caddy2
had multiple issues related to GPG key signing and couldn't update:
https://aur.archlinux.org/packages/caddy2/
https://github.com/caddyserver/caddy/issues/3797

As a result I'm copying some of their caddy service file. I don't know
what all of these options do, it's a bit of cargo culting I guess.

Also, to prevent browsers from attempting to fetch a favicon that does
not exist, we're adding a blank favicon to the html of the test page:
https://stackoverflow.com/a/38917888/7921798