generate Hugo websites on server
I scrubbed extra images from the git history using git-filter-repo:
https://github.com/newren/git-filter-repo
Then I created a new repo so that we wouldn't accidentally re-add
the bad old history. I used this command:
git filter-repo --path public --path resources --invert-paths

Now we can't just copy the git repo that has the images in it, we must
generate the images on the server by running the Hugo command, as we
did in the original Sunrise role. The good news is that since we are
using Hugo modules for the theme, we only need to watch the content
repo for changes. We update the theme by updating the theme module
version number in go.mod, and changing go.mod will count as changing
the content directory.