update Caddyfile format, test RI site, permissions
The RI site appears to be running fine with local certs on america,
we should be able to point the domain at america now and use public
certs.

PHP wasn't working because the caddy user and the PHP user "http" were
different, so Caddy didn't have permissions to interact with PHP. We
fixed this by adding the caddy user to the http group.

Another bug that we solved manually was the Caddy service is not being
restarted after the binary is updated. We still need to trigger a
restart when the package manager updates Caddy in this role.