pass Mozilla Observatory tests, separate css
On our test website, we got to an A+, 120/100!
https://observatory.mozilla.org/analyze/test.maximumethics.dev

Turns out all .dev domains are automatically on the HSTS preload list,
so we don't have to do anything for that security measure yet:
https://endtimes.dev/dev-and-hsts-preload/
https://get.dev/

Content-Security-Policy header blocks inline CSS from working, so css
needs to go in its own file, and then explicitly allowed.
https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP

My attempt to prevent the browser from loading a non-existent favicon
requires us to explicitly allow a data url:
https://webdesign.tutsplus.com/tutorials/prevent-automatic-favicon-requests--cms-34762
Probably we should just get a favicon, but it's annoying how difficult
it is to disable browsers requesting a favicon if you really don't want
one.

I wanted to only provide the CSS file as an HTTP header:
https://meiert.com/en/blog/prefer-http-headers/
But aside from it only working in Firefox, Firefox's dev tools seem to
be unable to refresh updates to the CSS when delivered this way, even
though the browser displays CSS updates just fine. Since one of the top
reasons I would want to do this is to tinker with alternate CSS files, I
gave up on the idea.

I'm committing to www. as the canonical URL for all of our websites:
https://www.yes-www.org/
Since I intend to support other protocols such as Gemini, I think it's
worth drawing the distinction with the Web.
Also, until SRV records or some other solution for locating HTTP servers
is standardized, I'm convinced we don't want naked domains for technical
reasons.
https://stackoverflow.com/questions/9063378/why-do-browsers-not-use-srv-records