remove support for AUR
Caddy is getting an official Arch package again, so we can ditch AUR.
Caddy is currently in Community-Testing:
https://archlinux.org/packages/community-testing/x86_64/caddy/
I manually enabled that repo:
https://joshtronic.com/2019/08/19/how-to-install-packages-from-testing-on-arch-linux/
If it doesn't come out of testing soon we should write a play to do this

I also manually removed the AUR packages and the aur_builder user.

Bumped Hugo version to v0.81.0 on Ubuntu.

Implemented possibly brittle Content-Security-Policy for
maximumethics.dev, but we can fix it when we inevitably trip over it.

Converted CSS whitespace to tabs, but YAML doesn't like that so can't do
that with Ansible playbooks. Began implementing fallbacks for NetSurf.