put new servers in "new" hosts group
new-linode.yml currently cannot run on established servers that don't
use the default credentials. So we don't want to wait around for those
established servers to time out. Instead we only want to run the
server creation tasks on just-created servers. So let's create a "new"
group to hold them.