don't set up user shells on server creation
This is not intended to set up a usable environment, just the bare
minimum to turn on a server. Therefore we can delay giving users their
preferred shells until the primary role (currently called "sunrise")
runs.

Also, Ansible lies about setting up user shells on Alpine Linux, so
this does nothing anyway, but that's a problem for another day.
https://github.com/ansible/ansible/issues/61911