initial support for Alpine Linux and apk packages
Alpine may have the fastest package manager:
https://michael.stapelberg.ch/posts/2019-08-17-linux-package-managers-are-slow/

It is famously small and minimalist, good for conserving resources. It
uses alternative software by e.g. eschewing systemd and coreutils,
promoting software diversity and resilience. In short, it seems like a
good fit for our ethos.

Personally I also prefer rolling distros, so I like that they have the
"edge" rolling distro option, but annoyingly Linode does not offer an
edge image, so we'd have to create one ourselves. For now we just use
the latest stable version, 3.14.

Two large (and related) problems with Alpine are that (1) their non-
standard software choices mean that you run into strange edge cases
when bootstrapping that are not well-worn paths, and (2) most people
who run Alpine run it in a container. This means that we are
constantly discovering bugs, lack of support, and random footguns that
other people rarely encounter, and which are not an issue with other
Linux distros.

Still, Alpine is simple enough that once we get it working, I suspect
it may be easier to maintain in some ways than other distros.