caddy: configure Caddy + php-fpm on Alpine Linux
php-fpm on Alpine Linux does not default to using a Unix socket. I
don't know enough to have an opinion on whether this is good or not,
but it seems good to support this option, even if technically I could
simply change the php-fpm config to use a Unix socket, as Arch and
Ubuntu do by default.