serve static archives of Wordpress sites
Wordpress is annoyingly complex to host. Hosting static files is much
simpler, and more similar to the Hugo hosting we're already doing.

Therefore, rather than spinning up actual Wordpress installs, for now
I'm going to use a plugin to make a static archive of each Wordpress
site and publish that:
https://wordpress.org/plugins/simply-static/

There are a number of downsides, including losing functions like
search (a feature of Simply Static's pro version).

Most annoyingly, this method relies on there being a copy of Wordpress
running somewhere. So, unless I commit to moving a website off of
Wordpress, and never making any more edits to the site in Wordpress,
I still either have to leave the old server running or install a real
copy of Wordpress on the new server. Or install Wordpress on each
local dev machine and run builds locally, or subscribe to some paid
static Wordpress hosting platform...